+++ packages/flutter_riverpod/test/src/core/provider_subscription_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -25,12 +27,12 @@ void main() {
     final container = tester.container();
     final navigator = tester.state<NavigatorState>(find.byType(Navigator));
 
-    navigator.pushNamed('/settings');
+    unawaited(navigator.pushNamed('/settings'));
     await tester.pumpAndSettle();
 
     container.read(dep.notifier).state++;
 
-    navigator.maybePop();
+    unawaited(navigator.maybePop());
     await tester.pumpAndSettle();
 
     expect(find.text('1'), findsOneWidget);